requirements.txt
Django==2.2.6
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
Markdown==2.6.11
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
sqlparse==0.3.0
